+++ src/Google/GoogleData.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace EventCalendar\Google;
+namespace Nexendrie\EventCalendar\Google;
 
-use EventCalendar\IEventModel;
+use Nexendrie\EventCalendar\IEventModel;
 use Nette\Utils\Arrays;
 
 /**