+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 0.7.x-dev
 - fixed default value for option wdayMaxLen
 - renamed package to nexendrie/event-calendar
 - BC break: removed BasicCalendar
+- BC break: changed namespace to Nexendrie\EventCalendar
 
 Version 0.6.1
 - initialized AbstractCalender::$onDateChange